async-api/requirements.txt
aioredis==2.0.1
elasticsearch[async]==8.5.1
fastapi==0.88
orjson==3.8.2
pydantic==1.10.2
uvicorn==0.20.0
uvloop==0.17.0
python-dotenv==0.21.0
fastapi-utils==0.2.1
fastapi-cache2==0.1.9
redis==4.3.4
black==22.10.0
gunicorn==20.1.0
dependency-injector==4.40.0